+++ Scripts/MMOGame/Database/SQLite/SQLiteDatabase_CharacterAttribute.cs
@@ -17,7 +17,7 @@ namespace Insthync.MMOG
             {
                 result = new CharacterAttribute();
                 result.dataId = reader.GetInt32("dataId");
-                result.amount = reader.GetInt32("amount");
+                result.amount = reader.GetInt16("amount");
                 return true;
             }
             result = CharacterAttribute.Empty;
