+++ Scripts/MMOGame/Database/SQLite/SQLiteDatabase_CharacterBuff.cs
@@ -20,7 +20,7 @@ namespace Insthync.MMOG
                 result.characterId = reader.GetInt64("characterId").ToString();
                 result.type = (BuffType)reader.GetByte("type");
                 result.dataId = reader.GetInt32("dataId");
-                result.level = reader.GetInt32("level");
+                result.level = reader.GetInt16("level");
                 result.buffRemainsDuration = reader.GetFloat("buffRemainsDuration");
                 return true;
             }
